style: Merged multiplier icons
